Fix SJS test linking
