+++ jvm/src/test/scala/org/scalamock/test/scalatest/AsyncMockFactoryNoDuplicatedRun.scala
@@ -1,15 +1,15 @@
 // Copyright (c) 2011-2015 ScalaMock Contributors (https://github.com/paulbutcher/ScalaMock/graphs/contributors)
-// 
+//
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
 // in the Software without restriction, including without limitation the rights
 // to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 // copies of the Software, and to permit persons to whom the Software is
 // furnished to do so, subject to the following conditions:
-// 
+//
 // The above copyright notice and this permission notice shall be included in
 // all copies or substantial portions of the Software.
-// 
+//
 // THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 // IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 // FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -18,35 +18,42 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package org.scalamock.test.specs2
+package org.scalamock.test.scalatest
 
-import org.scalamock.test.mockable.TestTrait
-import org.scalamock.specs2.IsolatedMockFactory
-import org.specs2.mutable.Specification
+import org.scalamock.scalatest.AsyncMockFactory
+import org.scalatest.flatspec.AsyncFlatSpec
+import org.scalatest.matchers.should.Matchers
+
+import scala.concurrent.Future
 
 /**
- *  Tests for mocks defined in suite scope (i.e. outside test case scope)
- *
- *  Tests for issue #25
- */
-class SuiteScopeMockParallelTest extends Specification with IsolatedMockFactory {
-  // please note that this test suite runs in isolated mode
+  * Test to ensure AsyncMockFactory only run test once
+  */
+class AsyncMockFactoryNoDuplicatedRun extends AsyncFlatSpec with Matchers with AsyncMockFactory {
+  trait TestTrait {
+    def mockMethod(): Int
+  }
 
-  val mockWithoutExpectationsPredefined = mock[TestTrait]
-
-  "Specs2 suite" should {
-    "allow to use mock defined suite scope" in {
-      (mockWithoutExpectationsPredefined.oneParamMethod _).expects(1).returning("one")
-      (mockWithoutExpectationsPredefined.oneParamMethod _).expects(2).returning("two")
-
-      mockWithoutExpectationsPredefined.oneParamMethod(1) must_== "one"
-      mockWithoutExpectationsPredefined.oneParamMethod(2) must_== "two"
+  class ClassUnderTest(protected val testTrait: TestTrait) {
+    def methodUnderTest(): Future[Int] = {
+      TestCounter.alreadyRun = TestCounter.alreadyRun + 1
+      Future(testTrait.mockMethod())
     }
+  }
 
-    "allow to use mock defined suite scope in more than one test case" in {
-      (mockWithoutExpectationsPredefined.oneParamMethod _).expects(3).returning("three")
+  object TestCounter {
+    var alreadyRun: Int = 0
+  }
 
-      mockWithoutExpectationsPredefined.oneParamMethod(3) must_== "three"
-    }
+  "AsyncMockFactory" should "run test case provided successfully" in {
+    val mockTrait = mock[TestTrait]
+    val returnVal = 100
+    (() => mockTrait.mockMethod()).expects().returning(returnVal)
+    val classUnderTest = new ClassUnderTest(mockTrait)
+    classUnderTest.methodUnderTest().map(_ shouldBe returnVal)
+  }
+
+  "AsyncMockFactory" should "have run test case provided only once" in {
+    Future(TestCounter.alreadyRun shouldBe 1)
   }
 }
