+++ jvm/src/test/scala/org/scalamock/test/scalatest/MixedMockFactoryTest.scala
