+++ jvm/src/test/scala/org/scalamock/test/scalatest/SuiteScopeProxyMockTest.scala
